apps/agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "awaaz-agent"
version = "0.1.0"
description = "Awaaz LiveKit voice agent worker (secondary channel)"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Gelecek Solution", email = "engineering@awaaz.pk" }]
dependencies = [
    "livekit-agents[deepgram,anthropic,openai,silero,turn-detector]>=1.5.6",
    "livekit-plugins-upliftai>=1.2.16",
    "anthropic>=0.40",
    "deepgram-sdk>=4.0",
    "httpx>=0.27",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0.35",
    "structlog>=24.4",
    "twilio>=9.3",
    "phonenumbers>=8.13",
    "dateparser>=1.2",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp>=1.27",
    "tenacity>=9.0",
    "orjson>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.6.9",
    "black>=24.10",
    "mypy>=1.11",
]
local = [
    "faster-whisper>=1.0",
    "ctranslate2>=4.5",
    "transformers>=4.45",
    "piper-tts>=1.2",
    "onnxruntime>=1.18",
]

[tool.setuptools]
packages = ["awaaz_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "ASYNC", "S", "RUF", "TID"]
ignore = ["S101", "B008"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["livekit.*", "twilio.*", "deepgram.*", "faster_whisper.*", "transformers.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["awaaz_agent/tests"]
markers = [
    "integration: requires real provider credentials",
    "eval: end-to-end LLM evaluation suite (slow)",
]
